+site

+++ src/main/java/com/frequency/frequencyplanner/domain/FrequencyBand.java
@@ -8,7 +8,7 @@ public class FrequencyBand {
     	this.frequency=frequency;
     }
 
-	public long getFrequency() {
+	public int getFrequency() {
         return frequency;
     }
